feat(api,async): reorganising code
